Migration of ovn/services.go into the node package
This removes service control functionality from the ovn package and
merges it into the node package alongside similar service related
functionality.

This is the second commit in regards to debloating the ovn package in an
attempt to remove unnecessary functionality and dependencies from the codebase
